seperate APIRequest code to new file
