+++ draft/draft/Models/RoomsByDate.swift
@@ -31,13 +31,8 @@ class RoomsByDate: Sequence, IteratorProtocol {
         addRoom(room: room)
     }
     
-    // add room test function
     func addRoom(room: Room) {
         rooms.append(room)
-    }
-    
-    func getDateInString() -> String {
-        return date.dateToString
     }
     
     // implement IteratorProtocol protocols to use 'for in RoomGroup'
@@ -52,9 +47,5 @@ class RoomsByDate: Sequence, IteratorProtocol {
         }
     }
 }
-
-// sample data
-//var sampleRoom = Room(name: "집 앞 농구장에서 한 게임 해요" , date: Date(), maxNum: 4)
-
